Add "migration of the objects" to Fork section
This pull request is not all of the contents of `4-migration-of-the-objects.md`. To make easier to review and modify, I will submit the rest of contents dividedly. Because I think this content is too long to review and modify in a single pull request. When this pull request is reviewed and I reflect suggestions, I will submmit next pull request.